2026年3月4日 下午5:14

+++ UVa/00534.cpp
@@ -3,7 +3,7 @@
 #define pll pair<long,long>
 #define ll long long
 using namespace std;
-int M[201][201];
+ll M[201][201];
 int main()
 {
     ios::sync_with_stdio(0), cin.tie(0), cout.tie(0);
